Added footprint for the Hope RFM69HW
Both the SMD configuration and the header configuration
for the RFM69HW module is supported in this footprint.

No 3D model is referenced due to missing suitable components
in the standard KiCad repositories. However, a 2mm pitch header
can be found here for those who want to add at least the headers
as a 3D model for this part:
http://smisioto.no-ip.org/elettronica/kicad/kicad-en.htm
git://smisioto.eu/kicad_libs.git